Order details: fixed order operations disappearing after doing any update that causes totals recalculation (e.g. updating money amounts for any payment).
